requirements.txt
telebot~=0.0.5
numpy~=1.18.5
pillow~=7.2.0
torch~=1.13.1
opencv-python~=4.5.2.52
aiohttp
asyncio